wasp_agents/mock_agents/requirements.txt
a2a-sdk>=0.2.6
openai>=1.0.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0 